Restructure "bad combinations"
Passwords not matching is now solely the responsibility of the
confirmPassword field. However, I still want to show off that
Form.Result can produce any kind of error type, so split up the
passwordNotStrongEnough into passwordNoLetter, passwordNoNumber, and
passwordNoSymbol.

Fractions not being proper is now the responsibility of a new
"combined" field in FractionError. While we're here, add a new error
case for negative numbers in both numerator and denominator.