Separate "andErr" into "maybeErr" and "unconditionalErr"
unconditionalErr is a little dangerous because it fills out an error
field but since it can't check whether there was an error, it cannot
update the form state. Renaming it to "unconditional" makes it sound a
little more serious. "andErr" is a kind of neutral name, so rename it
to "maybeErr" by analogy with "maybeValid".