Turns out we don't need ifNothing
Maybe we shouldn't include it in this library. It may be the case that
it's only useful in my own projects, in which case I should just move
it there.

Ideally we could get it included in maybe-extra but that seems
unlikely.

Let's leave it in this project for now.