Provide a better solution for "bad combination" in the example
`makePair` is infallible, so the old example doesn't do a great job of
demonstrating a "bad combination" problem, where you might not even be
able to put together a combination if the combination is
bad (i.e. where your `makePair` solution returns a `Maybe` or
`Result`). To make a more realistic example, create a `makeFraction :
Int -> Int -> Maybe Fraction` and use it in the example.